refactor: split rendering data preparation into focused modules
Keep card states, form display models, feedback, progress, navigation, and page content in their own rendering modules. Preserve the submission contract and learner-visible behavior while updating callers, regression coverage, and contributor documentation.
